refactor: move presenters into lib
