+++ gapic-generator/templates/default/helpers/filepath_helper.rb
@@ -14,25 +14,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-require "active_support/inflector"
+require "gapic/helpers/filepath_helper"
 
-module FilepathHelper
-  ##
-  # Converts a ruby namespace string to a file path string.
-  def ruby_file_path api, namespace
-    file_path = ruby_file_path_for_namespace namespace
-    fix_file_path api, file_path
-  end
-
-  ##
-  # Converts a ruby namespace string to a file path string.
-  def ruby_file_path_for_namespace namespace
-    ActiveSupport::Inflector.underscore namespace
-  end
-
-  ##
-  # Corrects a namespace by replacing known bad values with good values.
-  def fix_file_path api, file_path
-    file_path.split("/").map { |node| api.fix_file_path node }.join("/")
-  end
-end
+FilepathHelper = Gapic::Helpers::FilepathHelper
