+++ gapic-generator/templates/default/helpers/namespace_helper.rb
@@ -14,31 +14,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-require "active_support/inflector"
+require "gapic/helpers/namespace_helper"
 
-module NamespaceHelper
-  ##
-  # Looks up the ruby_package for a dot-separated address string to a new string
-  # and creates the corrected Ruby namespace
-  def ruby_namespace api, address
-    file = api.file_for address
-    address = address.dup
-    address[file.package] = file.ruby_package if file.ruby_package.present?
-    namespace = ruby_namespace_for_address address
-    fix_namespace api, namespace
-  end
-
-  ##
-  # Converts an array or dot-separated address string to a new string with
-  # Ruby double-semicolon separators.
-  def ruby_namespace_for_address address
-    address = address.split "." if address.is_a? String
-    address.reject(&:empty?).map(&:camelize).join "::"
-  end
-
-  ##
-  # Corrects a namespace by replacing known bad values with good values.
-  def fix_namespace api, namespace
-    namespace.split("::").map { |node| api.fix_namespace node }.join("::")
-  end
-end
+NamespaceHelper = Gapic::Helpers::NamespaceHelper
