--- gapic-generator/templates/default/helpers/presenter_helper.rb
@@ -1,24 +1,0 @@
-# frozen_string_literal: true
-
-# Copyright 2018 Google LLC
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     https://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-require "active_support/inflector"
-require_relative "presenters/gem_presenter"
-
-module PresenterHelper
-  def gem_presenter api
-    GemPresenter.new api
-  end
-end
