+++ gapic-generator/test/gapic/presenters/gem/garbage_test.rb
@@ -18,7 +18,7 @@ require "test_helper"
 
 class GarbageGemPresenterTest < PresenterTest
   def presenter
-    GemPresenter.new api :garbage
+    Gapic::Presenters::GemPresenter.new api :garbage
   end
 
   def test_garbage
@@ -38,13 +38,13 @@ class GarbageGemPresenterTest < PresenterTest
     assert_equal "GARBAGE", presenter.env_prefix
 
     assert_equal ["endless.trash.forever"], presenter.packages.map(&:name)
-    presenter.packages.each { |pp| assert_kind_of PackagePresenter, pp }
+    presenter.packages.each { |pp| assert_kind_of Gapic::Presenters::PackagePresenter, pp }
 
     assert_equal ["GarbageService"], presenter.services.map(&:name)
-    presenter.services.each { |sp| assert_kind_of ServicePresenter, sp }
+    presenter.services.each { |sp| assert_kind_of Gapic::Presenters::ServicePresenter, sp }
 
     assert_equal ["google/api/field_behavior.proto", "google/api/resource.proto", "google/protobuf/any.proto", "google/protobuf/empty.proto", "google/rpc/status.proto", "google/longrunning/operations.proto", "google/protobuf/timestamp.proto", "google/protobuf/duration.proto", "google/iam/v1/policy.proto", "google/iam/v1/iam_policy.proto", "garbage/garbage.proto"], presenter.proto_files.map(&:name)
-    presenter.proto_files.each { |fp| assert_kind_of FilePresenter, fp }
+    presenter.proto_files.each { |fp| assert_kind_of Gapic::Presenters::FilePresenter, fp }
 
     assert presenter.iam_dependency?
   end
