+++ gapic-generator/test/gapic/presenters/service/showcase_echo_test.rb
@@ -27,7 +27,7 @@ class ShowcaseEchoServiceTest < PresenterTest
 
   def test_methods
     refute_empty presenter.methods
-    presenter.methods.each { |ref| assert_kind_of MethodPresenter, ref }
+    presenter.methods.each { |ref| assert_kind_of Gapic::Presenters::MethodPresenter, ref }
     exp_method_names = ["echo", "expand", "collect", "chat", "paged_expand", "wait", "block"]
     assert_equal exp_method_names, presenter.methods.map(&:name)
   end
@@ -137,7 +137,7 @@ class ShowcaseEchoServiceTest < PresenterTest
   end
 
   def test_lro_service
-    assert_kind_of ServicePresenter, presenter.lro_service
+    assert_kind_of Gapic::Presenters::ServicePresenter, presenter.lro_service
   end
 
   def test_lro_client_var
